fix duty cycle not changing
forgot OCR1A was already a pointer, and to remove param declars

+++ PWM.hpp
@@ -5,8 +5,8 @@
 
 void increaseDutyCycle(int* OCR1A_reg);
 void dereaseDutyCycle(int* OCR1A_reg);
-void increaseDutyCycleBy(float delta, int *OCR1A_reg);
-void decreaseDutyCycleBy(float delta, int *OCR1A_reg);
+void increaseDutyCycleBy(int delta, int *OCR1A_reg);
+void decreaseDutyCycleBy(int delta, int *OCR1A_reg);
 void setDutyCycle(int* OCR1A_reg, DutyCycle_t dutyCycle);
 
 #endif
